example-skills/code-atlas/pyproject.toml
[project]
name = "code-atlas"
version = "0.1.0"
description = "Agent-oriented Python codebase structure and metrics analyzer"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "DotWork Team", email = "team@dotwork.dev" }
]
license = { text = "MIT" }
dependencies = [
    "typer>=0.9.0",
    "radon>=6.0.0",
    "pyyaml>=6.0",
    "watchdog>=3.0.0",
    "ruff>=0.1.9",
    "mypy>=1.7.1",
    "types-pyyaml>=6.0.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "pytest-timeout>=2.2.0",
    "pytest-asyncio>=0.21.1",
]
visual = [
    "rich>=13.7.0",
    "networkx>=3.2",
    "pyvis>=0.3.2",
]

[project.scripts]
code-atlas = "code_atlas.cli:app"

[project.entry-points."glorious_agents.skills"]
atlas = "code_atlas.skill:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/code_atlas"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--cov=src/code_atlas",
    "--cov-report=term-missing",
    "--cov-report=xml",
    "--cov-fail-under=70",
    "--timeout=5",
    "-v",
]
timeout = 5

[tool.coverage.run]
source = ["src/code_atlas"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/.venv/*",
    "*/cli.py",
    "*/cli_old.py",
    "*/scanner_old.py",
    "*/watch_commands.py",
    "*/daemon_utils.py",
    "*/analysis_commands.py",
    "*/ast_extractor.py",
    "*/metrics.py",
    "*/git_analyzer.py",
    "*/dependency_graph.py"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "S",   # flake8-bandit (security)
    "B",   # flake8-bugbear
]
ignore = [
    "E501",  # Line too long (handled by formatter)
    "S603",  # subprocess call without shell=True
    "S607",  # Starting process with partial path (git is trusted)
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_unimported = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
check_untyped_defs = true
strict_equality = true

[[tool.mypy.overrides]]
module = "networkx.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pyvis.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "watchdog.*"
ignore_missing_imports = true

[dependency-groups]
dev = [
    "types-psutil>=7.0.0.20251111",
]
